requirements.txt
flask>=2.3.0,<4.0.0
numpy>=1.24.0,<2.0.0
joblib>=1.3.0,<2.0.0
fasttext-wheel>=0.9.2,<1.0.0
scikit-learn>=1.3.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
huggingface_hub>=0.20.0,<1.0.0
